move OwnCloudClient redirect functionallity into redirection manager
